# Slight simplification of chars().count()