@arniu 2021-01-26
How about `self.iter.filter(|byte| !utf8_is_cont_byte(byte)).count()`?