@SkiFire13 2021-01-26
It doesn't look like this makes any difference https://godbolt.org/z/vKhTo9
The only difference I can see is that there are two additional `mov`s and two additional `sub`s, so a tiny regression.

> Is there a github command to kick off a performance comparison?

There's one to profile the compiler, but I don't this is the case. Usually you just locally run the relevant benchmarks and post the results. 

ps: Does anyone know why the related benches are in `library/alloc/benches/str.rs`?