@bugadani 2021-01-26
> The only difference I can see is that there are two additional movs and two additional subs, so a tiny regression.

I think you're looking at the diff backwards, Editor #2 seems to be the smaller one. Not a terribly big difference, though. Either way the code requires a bit of a "what's going on here?" thinking, but I find the shorter version easier on my brain.

> How about `self.iter.filter(|byte| !utf8_is_cont_byte(byte)).count()`?

Even better IMO.