@bluss 2021-01-26
Nice simplification!