@gilescope 2021-01-26
Godbolt has nice diffs! It shows that filter count is exactly the same instructions as map sum so I'll change it to filter count.