@rust-log-analyzer 2021-01-26
The job **`x86_64-gnu-llvm-9`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/runs/1772637562) [(plain)](https://github.com/rust-lang/rust/commit/d0eb2d6615a8f46deae1d8822fc3c4aa361cbb2b/checks/1772637562/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
   Compiling unwind v0.0.0 (/checkout/library/unwind)
error[E0308]: mismatched types
  --> library/core/src/str/iter.rs:50:53
   |
50 |         self.iter.filter(|&byte| !utf8_is_cont_byte(byte)).count()
   |                                                     |
   |                                                     |
   |                                                     expected `u8`, found `&u8`
   |                                                     help: consider dereferencing the borrow: `*byte`
error: aborting due to previous error

For more information about this error, try `rustc --explain E0308`.
error: could not compile `core`
```

</details>