@gilescope 2021-01-26
(@SkiFire13 thanks for pointing out where the benchmarks were.)
Running `./x.py bench library/alloc --test-args chars_count`

Before:
test str::chars_count::long_lorem_ipsum                  ... bench:         734 ns/iter (+/- 134)
test str::chars_count::short_ascii                       ... bench:          19 ns/iter (+/- 3)
test str::chars_count::short_mixed                       ... bench:          20 ns/iter (+/- 5)
test str::chars_count::short_pile_of_poo                 ... bench:          19 ns/iter (+/- 4)

After:
test str::chars_count::long_lorem_ipsum                  ... bench:         730 ns/iter (+/- 93)
test str::chars_count::short_ascii                       ... bench:          18 ns/iter (+/- 2)
test str::chars_count::short_mixed                       ... bench:          19 ns/iter (+/- 2)
test str::chars_count::short_pile_of_poo                 ... bench:          19 ns/iter (+/- 2)

So seems no noticable difference in performance which given it only shaves a few instructions off I guess is to be expected.