@klensy 2021-01-30
```rust
fn count(self) -> usize {
    self.iter.filter(|&&byte| !utf8_is_cont_byte(byte)).take(usize::MAX).count()
}
```

This somehow generates smaller code, but maybe it slower?