@SkiFire13 2021-01-30
That doesn't vectorize the loop so yes, it would probably be slower